// components/scout/types.ts
export enum ScoutSwitchType {
  index = "Index",
  quickBridge = "QuickBridge",
  bidder = "Bidder",
}

export enum QuickBridgeTabType {
  vehile = "Vehicle",
  oem = "OEM",
  class = "Class",
  segment = "Segment",
  technology = "Technology",
  commodity = "Commodity",
  productionTech = "Production Tech",
  services = "3P Services",
  pioneer = "Pioneer",
}